+++ CHANGELOG.md
@@ -8,6 +8,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ### Added
+- Text and JSON handlers compatible with slog.
+- Support for reopening log files.
+
+### Changed
+- Logger outputs close is thread-safe now.
+
+### Fixed
+
+
+## [v1.0.0] - 2026-01-13
+
+### Added
 
 - Core structured logging library for Go.
 - Support for log levels: Trace, Debug, Info, Warn, Error.
@@ -18,7 +30,3 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Test suite for core functionality.
 - Idiomatic Go examples (testable examples).
 - Makefile, GitHub Actions CI workflow, README, LICENSE, lint configuration.
-
-### Changed
-
-### Fixed